fix: update loop
